requirements.txt
torch
pandas
sentence_transformers
faiss-cpu